Refactored code related to products filtering feature.
