+++ src/Plugin.php
@@ -43,13 +43,17 @@ class Plugin {
 
     add_action('parse_request', __CLASS__  . '::parse_request');
 
-    // We need woocommerce-german-market plugin being active to add product
-    // delivery time support to products filtering in the woocommerce layered
-    // navigation.
+    // Allow to filter products by delivery time if woocommerce-german-market plugin is active.
     $active_plugins = apply_filters('active_plugins', get_option('active_plugins'));
     if (in_array('woocommerce-german-market/WooCommerce-German-Market.php', $active_plugins)) {
       // Registers products filter widgets supporting delivery time.
-      add_action('widgets_init', __CLASS__ . '::widgets_init');
+      add_action('widgets_init', __NAMESPACE__ . '\ProductFilters\DeliveryTime::widgets_init');
+
+      if (!is_admin()) {
+        // Adds custom query variable to filter products by delivery time.
+        add_filter('query_vars', __NAMESPACE__ . '\ProductFilters\DeliveryTime::query_vars');
+        add_action('pre_get_posts', __NAMESPACE__ . '\ProductFilters\DeliveryTime::pre_get_posts', 1);
+      }
     }
   }
 
@@ -126,10 +130,6 @@ class Plugin {
       remove_filter('woocommerce_order_item_name', ['WGM_Template', 'add_delivery_time_to_product_title']);
     }
 
-    // Adds custom query variable to filter products by delivery time.
-    add_filter('query_vars', __NAMESPACE__ . '\WooCommerce::query_vars');
-    add_action('pre_get_posts', __NAMESPACE__ . '\WooCommerce::pre_get_posts', 1);
-
     // Enqueues plugin scripts.
     add_action('wp_enqueue_scripts', __CLASS__ . '::wp_enqueue_scripts');
   }
@@ -153,24 +153,6 @@ class Plugin {
       wp_redirect(home_url($shop_page_slug), 301);
       exit;
     };
-  }
-
-  /**
-   * Registers products filter widgets supporting delivery time.
-   *
-   * @implemets widgets_init
-   */
-  public static function widgets_init() {
-    // Registers widget to filter products by delivery time.
-    register_widget(__NAMESPACE__ . '\Widgets\WidgetFilterDeliveryTime');
-
-    // Overrides layered nav woocommerce widgets with new ones supporting
-    // woocommerce-german-market delivery time taxonomy terms to be used
-    // as product filters.
-    unregister_widget('WC_Widget_Layered_Nav_Filters');
-    unregister_widget('WC_Widget_Layered_Nav');
-    register_widget(__NAMESPACE__ . '\Widgets\WidgetLayeredNav');
-    register_widget(__NAMESPACE__ . '\Widgets\WidgetLayeredNavFilters');
   }
 
   /**
@@ -247,33 +229,6 @@ class Plugin {
    */
   public static function wpseo_disable_adjacent_rel_links() {
     return get_option(Plugin::L10N . '_wpseo_disable_adjacent_rel_links');
-  }
-
-  /**
-   * Adds the passed argument as query parameter to all matched hrefs.
-   *
-   * @param string $html_filter
-   *   The content to perform the transformation on.
-   * @param string $filter_name
-   *   The query parameter to add.
-   *
-   * @return string
-   */
-  public static function addFilterToNavLinks(string $html_filter, string $filter_name): string {
-    if ($filter_args = $_GET[$filter_name] ?? []) {
-      $filter_args = array_filter(array_map('absint', explode(',', wp_unslash($filter_args))));
-    }
-    // Return early if filter is currently not active.
-    if (!$filter_args) {
-      return $html_filter;
-    }
-    // Add query parameter to all found hrefs.
-    $html_filter = preg_replace_callback('@href="(.+?[^"])"@', function ($match) use ($filter_name, $filter_args) {
-      $link = 'href="' . esc_url(add_query_arg($filter_name, implode(',', $filter_args), $match[1])) . '"';
-      return $link;
-    }, $html_filter);
-
-    return $html_filter;
   }
 
   /**
